backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0
pymongo==4.5.0
python-dotenv==1.0.0
bcrypt==4.0.1
email-validator==2.0.0
python-jose==3.3.0